Code Samples from Independent Prototyping Project
while learning Unity, C#, and moving through the unity Learn Platform I created a multiple level, multiple game mode prototype based around the concept of hitting a target with several control schemes. Code is rough looking back at it, but does its job. Leaving in current state to show a line of progress as well as proving the writing of original code (with reference of course *noted when important*)  as opposed to blindly copying it for results. I believe anyone can be trained into good habits and a preferred layout for code, but it only matters if they can actually wright functional code themselves first.